data_loader/rgb_and_mask_renderer/requirements.txt
cycler==0.10.0
decorator==4.4.2
future==0.18.2
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.0
networkx==2.4
numpy==1.16.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.1
six==1.15.0
tifffile==2020.7.17
torch==1.10.1
torchvision==0.11.2
pycocotools==2.0.6
